{"schema":"decodex/commit/1","summary":"split prompting and landing gate modules","authority":"manual"}
